refactor: [#45] reorganize test fixtures
It does not make sense to have inependent mods and all the fixtures
together in the same file. I moved the fixtures to the mos that use
them.
