+++ tests/test_nautilus_librarian/test_mods/test_dvc/fixtures/dvc_fixtures.py
@@ -1,0 +1,7 @@
+import pytest
+
+
+@pytest.fixture()
+def temp_dvc_local_remote_storage_dir(tmp_path_factory):
+    fn = tmp_path_factory.mktemp("storage")
+    return fn
