+++ tests/test_nautilus_librarian/test_mods/test_shared/fixtures/shared_fixtures.py
@@ -1,0 +1,7 @@
+import pytest
+
+
+@pytest.fixture()
+def temp_dir(tmp_path_factory):
+    fn = tmp_path_factory.mktemp("test_dir")
+    return fn
